signals_ingestion/requirements.txt
pandas==0.23.4
pandas-gbq==0.9.0
asyncio
aiohttp
google-cloud-pubsub
prometheus_client
toolz